+++ src/generate/gen_aui_notebook.h
@@ -1,7 +1,7 @@
 //////////////////////////////////////////////////////////////////////////
 // Purpose:   wxAuiNotebook generator
 // Author:    Ralph Walden
-// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// Copyright: Copyright (c) 2020-2023 KeyWorks Software (Ralph Walden)
 // License:   Apache License -- see ../../LICENSE
 /////////////////////////////////////////////////////////////////////////////
 
@@ -15,9 +15,11 @@ class AuiNotebookGenerator : public BaseGenerator
 {
 public:
     wxObject* CreateMockup(Node* node, wxObject* parent) override;
+    void AfterCreation(wxObject* /*wxobject*/, wxWindow* /*wxparent*/, Node* /* node */, bool /* is_preview */) override;
 
     bool ConstructionCode(Code& code) override;
     bool SettingsCode(Code&) override;
+    bool AfterChildrenCode(Code&) override;
 
     bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
 
@@ -25,6 +27,8 @@ public:
     void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
 
     bool GetRubyImports(Node*, std::set<std::string>& /* set_imports */) override;
+
+    void AddPropsAndEvents(NodeDeclaration* declaration) override;
 
 protected:
     void OnPageChanged(wxBookCtrlEvent& event);